# swagger.json
{ "swagger": "2.0",
        "info": {
            "version": "1.0.0",
            "description": "API for E-commerce application",
            "title": "E-commerce API"
            },
        "host": "localhost:3000",
        "securityDefinitions": {
            "JWT": {
                "in": "header",
                "name": "Authorization",
                "type": "apiKey"
            }
        },

        "paths": {
            "/api/products": {
            "get": {
                "tags": ["Products"],
                "summary": "Get Products",
                "description": "User will get all products",
                "security": [{ "JWT": {} }],
                "responses": {
                    "200": {
                        "description": "OK"
                    },      
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
            },
            "/api/users/signin": {
                    "post": {
                    "tags": ["Users"],
                    "summary": "Login",
                    "description": "User login to get token",
                    "parameters": [
                    {
                    "in": "body",
                    "name": "body",
                    "description": "User Credentials",
                    "schema": {
                            "type": "object",
                            "properties": {
                            "email": {
                                "type": "string"
                            },
                            "password": {
                                "type": "string"
                            }
                        }
                    }
                    }
                    ],
                    "responses": {
                        "200": {
                            "description": "OK"
                        },
                        "400": {
                            "description": "Incorrect Credentials"
                        }
                    }
                }
            },

            "/api/cart":{
                "get": {
                    "tags":["Carts"],
                    "summary":"CARTITEM",
                    "description":"To get all the cartItems",
                    "security":[{"JWT":{}}],
        
                    "responses": {
                        "200": {
                            "description": "OK"
                        },
                        "400": {
                            "description": "Cart Item not found"
                        }
                    }


                        
                    

                }
                
            },
            "/api/cart/": {
                "post": {
                "tags": ["AddCart"],
                "summary": "Addcart",
                "description": "Add items in the cart",
                "security":[{"JWT":{}}],
                
                "parameters":[
                    {
                        "in":"query",
                        "name":"productId",
                       
                        "description":"plesse give product id",
                        "required":true
                    },
                    {
                        "in":"query",
                        "name":"quantity",
                        "description":"plesse give quantity of product",
                        "type":"integer"

                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item added"
                    },
                    "400": {
                        "description": "Incorrect Format"
                    }
                }
            }
        
            }
        }
}